requirements.txt
# Python dependencies for Flask app

Flask==2.0.2
google-cloud-firestore==2.3.2
google-cloud-storage==1.44.0
google-cloud-vision==3.5.0
matplotlib==3.5.1